+++ internal/module/delivery/shipping/mShipping/shipping-plan.go
@@ -19,10 +19,10 @@ type ShippingZoneFeeRule uint8
 
 // 运费规则类型定义
 var (
-	ShippingZoneFeeRuleOrderPrice   ShippingZoneFeeRule = 1 // 按订单价格
-	ShippingZoneFeeRuleProductPrice ShippingZoneFeeRule = 2 // 按商品价格
-	ShippingZoneFeeRuleProductCount ShippingZoneFeeRule = 3 // 按商品数量
-	ShippingZoneFeeRuleOrderWeight  ShippingZoneFeeRule = 4 // 按订单重量
+	ShippingZoneFeeRuleOrderPrice   ShippingZoneFeeRule = 1 // 按订单总价
+	ShippingZoneFeeRuleProductPrice ShippingZoneFeeRule = 2 // 按商品总价
+	ShippingZoneFeeRuleProductCount ShippingZoneFeeRule = 3 // 按商品件数
+	ShippingZoneFeeRuleOrderWeight  ShippingZoneFeeRule = 4 // 按包裹重量
 )
 
 // ShippingZoneFeeType 表示运费计算的类型
@@ -87,12 +87,12 @@ type ShippingZoneFee struct {
 // ShippingZonFeeCondition 表示物流区域方案运费的条件
 type ShippingZonFeeCondition struct {
 	mOrm.Model
-	Fixed             float64 `json:"fixed" gorm:"default:0"`            // 固定费用
-	First             float64 `json:"first" gorm:"default:0"`            // 首重/首件
-	FirstFee          float64 `json:"first_fee" gorm:"default:0"`        // 首重/首件的费用
-	Next              float64 `json:"next" gorm:"default:0"`             // 续重/续件
-	NextFee           float64 `json:"next_fee" gorm:"default:0"`         // 续重/续件的费用
-	Max               float64 `json:"max" gorm:"default:0"`              // 最大值限制
-	Min               float64 `json:"min" gorm:"default:0"`              // 最小值限制
+	Fixed             float32 `json:"fixed" gorm:"default:0"`            // 固定费用
+	First             float32 `json:"first" gorm:"default:0"`            // 首重/首件
+	FirstFee          float32 `json:"first_fee" gorm:"default:0"`        // 首重/首件的费用
+	Next              float32 `json:"next" gorm:"default:0"`             // 续重/续件
+	NextFee           float32 `json:"next_fee" gorm:"default:0"`         // 续重/续件的费用
+	Max               float32 `json:"max" gorm:"default:0"`              // 最大值限制
+	Min               float32 `json:"min" gorm:"default:0"`              // 最小值限制
 	ShippingZoneFeeId uint    `json:"shipping_zone_fee_id" gorm:"index"` // 关联物流区域方案运费ID
 }
